Move create_app to /warcor/init.py
