+++ warcor/models/campaign.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from warcor.app import db
+from warcor import db
 from warcor.utils.sanitisers import sanitise_input
